feat: course & course material resource controllers

+++ database/migrations/2024_08_10_100130_create_courses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->integer('duration');
+            $table->smallInteger('duration');
             $table->timestamps();
         });
     }
